Update docker-compose and configuration files; modify database port and service restart policy, adjust client repository search criteria, and enhance DTOs with new properties.

+++ src/modules/operacao/use-cases/pacote/delete-pacote/delete-pacote-use-case.ts
@@ -1,19 +1,29 @@
 import { inject, injectable } from 'tsyringe'
 import { Pacote } from '@modules/operacao/infra/typeorm/entities/pacote'
 import { IPacoteRepository } from '@modules/operacao/repositories/i-pacote-repository'
-import { HttpResponse } from '@shared/helpers'
+import { conflictError, HttpResponse } from '@shared/helpers'
+import { IEspelhoCargaItemsRepository } from '@modules/operacao/repositories/i-espelho-carga-items-repository'
 
 @injectable()
 class DeletePacoteUseCase {
-  constructor(@inject('PacoteRepository')
-    private pacoteRepository: IPacoteRepository
-  ) {}
+	constructor(
+		@inject('PacoteRepository')
+		private pacoteRepository: IPacoteRepository,
+		@inject('EspelhoCargaItemsRepository')
+		private espelhoCargaItemsRepository: IEspelhoCargaItemsRepository,
+	) {}
 
-  async execute(id: string): Promise<HttpResponse> {
-    const pacote = await this.pacoteRepository.delete(id)
+	async execute(id: string): Promise<HttpResponse> {
+		const alreadyHasEspelhoCarga = await this.espelhoCargaItemsRepository.getByPacoteItemId(id)
 
-    return pacote
-  }
+		if (alreadyHasEspelhoCarga.statusCode === 200) {
+			return conflictError('Pacote já tem espelho de carga')
+		}
+
+		const pacote = await this.pacoteRepository.delete(id)
+
+		return pacote
+	}
 }
 
 export { DeletePacoteUseCase }
