+++ src/modules/operacao/use-cases/pedido/delete-pedido/delete-pedido-use-case.ts
@@ -1,19 +1,29 @@
 import { inject, injectable } from 'tsyringe'
-import { Pedido } from '@modules/operacao/infra/typeorm/entities/pedido'
 import { IPedidoRepository } from '@modules/operacao/repositories/i-pedido-repository'
-import { HttpResponse } from '@shared/helpers'
+import { conflictError, HttpResponse } from '@shared/helpers'
+import { IPacoteRepository } from '@modules/operacao/repositories/i-pacote-repository'
+import { AppError } from '@shared/errors/app-error'
 
 @injectable()
 class DeletePedidoUseCase {
-  constructor(@inject('PedidoRepository')
-    private pedidoRepository: IPedidoRepository
-  ) {}
+	constructor(
+		@inject('PedidoRepository')
+		private pedidoRepository: IPedidoRepository,
+		@inject('PacoteRepository')
+		private pacoteRepository: IPacoteRepository,
+	) {}
 
-  async execute(id: string): Promise<HttpResponse> {
-    const pedido = await this.pedidoRepository.delete(id)
+	async execute(id: string): Promise<HttpResponse> {
+		const alreadyHasPacote = await this.pacoteRepository.getByPedidoId(id)
 
-    return pedido
-  }
+		if (alreadyHasPacote.statusCode === 200) {
+			throw new AppError('Pedido já tem pacote', 409)
+		}
+
+		const pedido = await this.pedidoRepository.delete(id)
+
+		return pedido
+	}
 }
 
 export { DeletePedidoUseCase }
